pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "generative-agents"
version = "1.0.0"
description = "A framework for generative multi-agent systems"
authors = [{ name = "Matúš Konig", email = "matuskonig@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "numpy>=2.2.0",
    "pydantic>=2.10.0",
    "openai>=1.68.0",
    'cachetools>=5.5.0',
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.2",
    "ipykernel>=6.21.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "networkx>=3.1.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["generative_multiagents"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line-length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
